+++ src/Libs/BluinoNet.ESP32StarterKit/SSD1306Imp.cs
@@ -35,7 +35,7 @@ namespace BluinoNet
         public Image CreateImage(
           int width,
           int height,
-          double[] data,
+          int[] data,
           int hScale,
           int vScale,
           Image.Transform transform)
